+++ lib/src/screens/recipe_import_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_translate/flutter_translate.dart';
 import 'package:nextcloud_cookbook_flutter/src/blocs/recipe/recipe.dart';
 import 'package:nextcloud_cookbook_flutter/src/screens/form/recipe_import_form.dart';
-import 'package:nextcloud_cookbook_flutter/src/screens/recipe_screen.dart';
+import 'package:nextcloud_cookbook_flutter/src/screens/recipe/recipe_screen.dart';
 
 class RecipeImportScreen extends StatelessWidget {
   final String importUrl;
